Order: Functions work, add username in lapangan database

+++ app/src/main/java/com/example/sportify/ui/ResultSearch.kt
@@ -143,7 +143,7 @@ fun FieldItem(field: Field, selectedDate: String, startTime: Int, endTime: Int, 
                                 selectedDate,
                                 startTime.toString(),
                                 endTime.toString(),
-                                field.price.toFloat()
+                                field.price.toDouble()
                             )
                         )
                     }
